// apps/buyer-keyword/src/app/shared/interfaces/buyer-payload.ts
export interface BuyerPayload {
  sourceSystemBuyerId: string;
  portalStatus: string;
  bounceBackMessage: string;
  forwardingEmailAddress: string;
  indexingSolutionId: string;
  displayPredictedValues: string;
  sourceSystemId: string;
  ingestionMethodId: string;
  buyerKeyword: string;
  displayIdentifierSearchValues: string;
}